requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
httpx==0.27.2
beautifulsoup4==4.12.3
apscheduler==3.10.4
pytest==8.3.3
email-validator==2.2.0
python-dotenv==1.0.1
